Implemented HATEOAS to reach Level 3 Maturity of REST API's.

+++ BiddingSystem/src/main/java/in/shuvam/service/ProductService.java
@@ -3,11 +3,11 @@ package in.shuvam.service;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import in.shuvam.entity.Products;
-import in.shuvam.exception.BidException;
-import in.shuvam.exception.ProductNotFound;
 import in.shuvam.repo.ProductRepo;
 
 @Service
@@ -27,23 +27,24 @@ public class ProductService {
 		return repo.findAll();
 	}
 
-	public String deleteProduct(int id) {
-		repo.deleteById(id);
-		return "Removed";
+	public void deleteProduct(int id) {
+		Products p = repo.findById(id).orElseThrow(()-> new ResponseStatusException(HttpStatus.NOT_FOUND,"Product not found"));
+		if(p != null)
+			repo.deleteById(id);
 	}
 
 	public Products getProduct(int id) {
-		return repo.findById(id).orElse(null);
+		return repo.findById(id).orElseThrow(()-> new ResponseStatusException(HttpStatus.NOT_FOUND,"Product not found"));
 		
 	}
 
 	public Products setBid(int id, double bid) throws Exception {
-		Products p = repo.findById(id).orElseThrow(()-> new ProductNotFound());
+		Products p = repo.findById(id).orElseThrow(()-> new ResponseStatusException(HttpStatus.NOT_FOUND,"Product not found"));
 		if (bid > p.getCurrent_bid()) {
 			p.setCurrent_bid(bid);
 			return repo.save(p);
 		} else
-			throw new BidException();
+			throw new ResponseStatusException(HttpStatus.NOT_ACCEPTABLE,"Bid must be higher than current bid");
 
 	}
